Fixed todos in jacobi test and pomp adapter
Subversion: revision:1875

+++ test/jacobi/hybrid/C++/main.mod.cpp
@@ -204,12 +204,11 @@ int pomp_num_threads = omp_get_max_threads();
     {
         for ( int i = 0; i < data.iCols; i++ )
         {
-            // TODO: check if this values have to be ints or doubles
-            int xx = ( int )( -1.0 + data.fDx * i );
-            int yy = ( int )( -1.0 + data.fDy * j );
+            double xx = ( -1.0 + data.fDx * i );
+            double yy = ( -1.0 + data.fDy * j );
 
-            int xx2 = xx * xx;
-            int yy2 = yy * yy;
+            double xx2 = xx * xx;
+            double yy2 = yy * yy;
 
             U( j, i ) = 0.0;
             F( j, i ) = -data.fAlpha * ( 1.0 - xx2 ) * ( 1.0 - yy2 )
